# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: APIs-Graduation
  version: 1.0.0
paths:
...
  /api/patient-group/patient/login:
    post:
      tags:
        - 'Patient Authentication'
      summary: 'Log in a patient'
      operationId: 06a2023b9b7548701ede2443f0d09239
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - email
                - password
              properties:
                email:
                  description: "Patient's email address"
                  type: string
                  format: email
                  example: patient@example.com
                password:
                  description: "Patient's password"
                  type: string
                  example: password123
              type: object
      responses:
        200:
          description: 'Login successful'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Login successful'
                  status:
                    type: integer
                    example: 200
                  token:
                    description: 'Sanctum authentication token'
                    type: string
                    example: 1|abc123xyz
                type: object
        401:
          description: 'Invalid credentials'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'The email is incorrect. Please verify that this email is correct.'
                  errors:
                    type: object
                  status:
                    type: integer
                    example: 401
                type: object
        403:
          description: 'Unauthorized action'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Unauthorized. Doctors cannot log in as patients.'
                  status:
                    type: integer
                    example: 403
                type: object
  /api/patient-group/patient/logout:
    post:
      tags:
        - 'Patient Authentication'
      summary: 'Log out a patient and deactivate their account'
      operationId: 4d9fa060124540111fa7339d5b619f28
      responses:
        200:
          description: 'Logged out and account deactivated successfully'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Logged out and account deactivated successfully'
                  status:
                    type: integer
                    example: 200
                type: object
        401:
          description: Unauthenticated
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: Unauthenticated
                  status:
                    type: integer
                    example: 401
                type: object
        403:
          description: 'Unauthorized action'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'You are not authorized to logout here.'
                  status:
                    type: integer
                    example: 403
                type: object
      security:
        -
          bearerAuth: []
...